Adding build script to update the Package.swift file
Creating update_swift_package.sh. This script relies on the OneSignal-XCFramework repo being in the same location as the OneSignal-iOS-SDK repo. The script zips the XCFramework in the OneSignal-iOS-SDK, computes its checksum (your command line must be using Swift tools 5.3 or later which comes with Xcode 12), writes the new checksum to the Package.swift file, then prompts for the new SDK version and updates the Package.swift file with the new version number. The zipped framework can then be dragged into the GitHub repo as a release asset